% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week4Project.R
\name{dt_trim}
\alias{dt_trim}
\title{Trim a data frame}
\usage{
dt_trim(X, mindate = NULL, maxdate = NULL)
}
\arguments{
\item{X}{is a time series containing a macroeconomic time series}

\item{mindate}{is a character variable representing the date the user wants the time series to start, if the data
allows this start date.}

\item{maxdate}{is a character variable representing the date the user wants the time series to end, if the data allows this end date.}
}
\value{
A dataframe is returned containing only data between the specified dates.
}
\description{
The data frame X is a time series data set with dates reaching back as far as the 1860s. This function
will trim that data set to the desired duration of time
}
