% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week4Project.R
\name{fleext}
\alias{fleext}
\title{Create an on line file extension}
\usage{
fleext(series, indx)
}
\arguments{
\item{series.}{Series is a reference to a macroeconomic data series. At this point one of the
series RGDP, CPI, Unemployment rate, or population.}

\item{indx.}{This is a number used to extract information about a call to the FRED. indx is determined
by the function fleindex.}
}
\value{
a url is returned that is used as the file extension to extract data from FRED.
}
\description{
This function is fed information from the main function and the fleindex function. The function
creates a url where macroeconomic data can be extracted from.
}
