% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week4Project.R
\name{fleindex}
\alias{fleindex}
\title{create a file index}
\usage{
fleindex(series)
}
\arguments{
\item{series.}{This input is controlled by other functions that call this function. It will be one of
three macroeconomic data series, RGDP, CPI, Unempmloyment rate, or Populaton.}
}
\value{
an integer is returned and is used to extract data in several other function calls.
}
\description{
This function assigns an index number to the series entered in the main function. This index number
is used serveral places to extract data for the call to FRED.
}
