% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week4Project.R
\name{groomdata}
\alias{groomdata}
\title{Create recession markers}
\usage{
groomdata(A, B, level)
}
\arguments{
\item{A}{is a dataframe of macroeconomic data created by the ME_dtaextract function and has be trimmed
using the ME_growth function (or not).}

\item{B}{is a dataframe created by the ME_Rec function.}

\item{level}{is a logical variable specifying if the level data should be matched to the recession data or if the
growth rate data should be matched to the recession data.}
}
\value{
A data frame of recession markers the same length as the input dataframe A is returned. This dataframe
is now ready to be plotted in the same.
}
\description{
This function requires two dataframes as input. With the input this function trims the B dataframe to the same
start date as the A dataframe.
}
