% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week4Project.R
\name{grth}
\alias{grth}
\title{Calculate a growth rate}
\usage{
grth(X)
}
\arguments{
\item{X}{is a data frame containing a time series of macroeconomic data.}
}
\value{
This functions returs a data frame containing the growth rates calculated from X
}
\description{
This function calculates a growth rate from an existing data series.
}
