% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week4Project.R
\name{ME_dtaextract}
\alias{ME_dtaextract}
\title{Extract Macroeconomic data}
\source{
St. Louis Federal Reserve Bank data (FRED)
}
\usage{
ME_dtaextract(series)
}
\arguments{
\item{series}{At this point series can be only one of four character strings, RGDP, Real Gross
domestic product, CPI, consumer price index, uneploymentrate, the US civilian unemployment rate,
and population, Total Population: All ages including Armed Forces Overseas.}
}
\value{
A dataframe is returned with a time-series for the given argument placed within the function.
This data frame is ready to be plotted and shown to a macroeconomic class, for example.
}
\description{
This function is given a single argument. The argement is a data macroeconomic data series
that is pulled from the federal reserve bank of St. Louis (FRED).
}
\examples{
ME_dtaextract("RGDP")
ME_dtaextract("Unemploymentrate")
ME_dtaextract("Population")


}
