% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week4Project.R
\name{ME_growth}
\alias{ME_growth}
\title{Calculate a growth rate}
\usage{
ME_growth(X, mindate = NULL, maxdate = NULL)
}
\arguments{
\item{X}{is a time series data frame. There should be a column containing a date class variable and a column
containing a macroeconomic variable.}

\item{mindate}{is a date in character fomat, in the form YYYY-mm-dd specifying the minimum date in the date set}

\item{maxdate}{is a date in character fomrate, in the form YYYY-mm-dd specifying the maximum date in the date set}
}
\value{
a dataframe is returned by this function containing two different growth rates, a growth
 rate per unit of time in the data and an annual growth rate.

 @import dplyr


 @examples ME_growth(X)
 @examples ME_growth(X, "1970-01-01")
 @examples ME_growth(X, "1965-09-07", "2010-01-01")


 @export
}
\description{
Time series data is frequently converted to growth rate data. In the conversion to growth rates
a data point within the data series is lost.
}
