% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Week4Project.R
\name{ME_series}
\alias{ME_series}
\title{Series test}
\usage{
ME_series(series)
}
\arguments{
\item{series}{This is the name of the macroeconomic data series to extract from FRED.
At this point series can take one of four sereis as an argument, RGDP (real Gross domestic product),
CPI (consumer price index), Unemploymentrate, (historical unemployment rates), Population,
this includes children and US service personel overseas, and Recession NBER dates for recessions.}
}
\value{
A data frame is returned with a macroeconomic series that is ready to plot.
}
\description{
This package has only four Macroeconomic data series currently being extracted. This functions checks if the chosen
series is available by this function.
}
\references{
This data is pulled from the Federal Reserve Bank of St. Louis (FRED).
}
